ggml-hexagon: release v1.06 and ready for code review
